Fixed some potentials segfaults, improved lisibility, optimized certain functions

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: jaubry-- <jaubry--@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/19 07:52:43 by jaubry--          #+#    #+#             */
-/*   Updated: 2024/11/05 17:01:21 by jaubry--         ###   ########.fr       */
+/*   Updated: 2024/11/14 09:54:33 by jaubry--         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	const unsigned char	*p1 = (const unsigned char *)s1;
 	const unsigned char	*p2 = (const unsigned char *)s2;
-	unsigned int		i;
+	size_t				i;
 
 	i = 0;
 	while (i < n)
